real_time_posenet_edge_tpu/edge_cloud_integration/cloud/src/convert/requirements.txt
google.cloud.storage
tensorflow==1.14.0